Changed direct constructor calls .new to .build, which provides a prettier wrapper without the expense of having to construct and deconstruct the argument list on #copy calls
Implemented #parent on most (or all) Schema abstract data types, which
may or may not be removed later. For the most part, the schema tree
structure will be mirrored by the value tree, but there are some
exceptions (HL and LS/LE loops) that make this potentially useful.

Used local variables scoped within a module for import-like
functionality, without re-exporting the imported constants. Too bad
local variable scopes within modules are not nested, but this is
consistent with Java's import mechanism -- which has benefits.

Moved Schema and Values classes for interchanges, functional groups, and
transaction sets into the Envelope namespace