Changed the semantics of Symbol#to_proc so [["a", "b"]].map(&:first) evaluates to ["a", "b"].first instead of "a".first("b"). Then #call can be used to evaluate [[:a, "b", "c"]].map{|m, *as| m.call(*as) } as "b".a("c"). This seems to resolve a fundamental difference in proc argument handling between new and old version of Ruby.
Updated AbstractState so we (naively) examine segment qualifiers -- it
works for now but can be optimized quite a bit by doing the work once
up-front instead of at each state.

Fixed a return value type error in InterchangeBuilder when it failed
to produce successor states.

Fixed the way in which tables are parsed. Like loops, the first segment
in the definition must occur for the table to occur. Unlike loops, if
the table doesn't have a header, the first segment can belong to a child
loop or even the table footer. I'm still not sure if this is correct,
but there is truly zero documentation on the syntax rules for tables.

Fixed a problem with how TableBuilder and TransactionSetBuilder
delegated to each other. Now when TransactionSetBuilder delegates down to
TableBuilder, TableBuilder knows not to delegate it back up again. When
TableBuilder delegates up to TransactionSetBuilder, TransactionSetBuilder
knows not to delegate it back down to *that* child, but it will correctly
delegate to other children.

Fixed a missing code for "40 - Receiver" in NM1's E98 qualifier element

Updated the test script to demonstrate parsing all the way to Loop
2010BB in an example file from Claredi. The file contains a syntax error
(apparently missing an HL segment to begin 2000C), and the parser reports
an unexpected CLM segment at the correct location.