Fixed the parser so all of the Claredi and X12 examples are parsed, with most of the changes centering around repeating the "Detail" tables.
Discovered highly annoying non-determinism when parsing a TableVal that
begins with a LoopVal -- occurrences of the loop start segment can mean
the loops repeated in the same table or in a new table. Implemented a
short-term workaround

Made significant optimizations after using the profiler. Not sure what
the previous performance was (given the bug fixes), but we can tokenize
and parse 6200 segments in 5 seconds. By changing the parser to build
state transition tables, we can still vastly improve upon this. Most of
the optimization came from preventing object cloning by destructively
reparenting from the root down only after we finished parsing the root.

Updated the pretty print methods to include the element names, which
should be verbose enough to make it *almost* readable by humans.