Implemented AbstractState#sequence, #tsequence, and #lsequence to generate state machine instruction sequences. Each concrete subclass of AbstractState now generates a list of instructions for parsing states that can directly succeed it.
Gained more insight to the "logic" of parsing TableDefs. Tables do
not need to begin with the first segment (eg PLB), and can begin with
any direct child segment in the definition OR with the start segment
of any direct child loop. This results in some nasty ambiguity when
both the table and the segment/loop can repeat -- does the next segment
belong to a new table or the existing one?

Implemented validation of the grammar to test for the above mentioned
ambiguity. Changed the 837P guide so it doesn't fail the test. This is
not a long-term solution -- the StateMachine will need to be modified
to identify this particular case of non-determinism and choose the
instruction that does not create a new table.