Resolved the push/pop problem by calculating the deepest level in the tree that a segment can belong from the list of instructions, and ignoring any instructions that would place the segment higher in the tree. This means as a general rule, any segments will be parented into the most nested structure rather than being pushed higher up the tree.
Stepped through some of the StateMachine's parsing algorithm and
identified and fixed a few bugs with InstructionTable. Renamed
Instruction#pop and #drop to #pop_count and #drop_count to avoid
confusion with the InstructionTable#pop and #drop methods.

Currently at a road block with TransactionSetState.push -- I haven't
found an easy way to push several layers of instructions onto the table
when transitioning from TransactionSetState to TableState or LoopState.
Need to devise a consistent way of managing the instruction stack when
an intermediate state does not consume a segment (eg ST), and ensure
that the last state we land on did consume it.