Created a parent exception class "StupiediError" and changed the existing exception classes to subclass it
Added BuilderDsl#successors so you can list the valid segments from the
parser's current state. Output shows the tree depth for each segment, etc.

Implemented FailureState, so unknown interchange, functional group, and
transaction set versions can be properly handled. Updated StateMachine
and BuilderDsl to throw a ParseError exception when an invalid segment
is generated programmatically. When reading from a file, the #read!
method does not throw an exception but leaves the StateMachine in a
stuck state and returns the unparsed input.

Fixed a minor formatting issue with the YARD abstract method plugin.