Updated the Generating.md document with more links to auto-generated documentation for several BuilderDsl methods
Improved documentation of BuilderDsl, added return type tags,
docstrings, etc

Made significant progress with the Navigation mixin. Changed the
return type of several methods to Either<StateMachine>. This is an
elegant solution because #node, #zipper, etc assume the machine is
in a deterministic state and can return Either.failure if that's not
the case.

Optimized Navigation methods by synchronizing the state and value
trees only at the stopping point states that are returned by the
method. The intermediate states shouldn't need to be synchronized,
but this needs to be carefully verified.

Implemented Navigation#find(segment_id) to efficiently search the
value tree based on information in the state tree. The potentially
reachable segments can be listed using #successors.

Renamed Navigation#forward and Navigation#backward to #next and #prev
to match Zipper::AbstractCursor.