Enhanced BuilderDsl to incrementally error-check constraints like element length, presence of required elements, absence of forbidden elements, etc.
Improved the error message for un-implemented abstract methods
to include the receiver's class name, instead of the class that
declared the abstract method

Fixed constructors for TimeVal::NonEmpty to ensure that #second
is a BigDecimal

Implemented SeparatorElementVal#too_long? and #too_short?. The
better solution would be not to make these elements required or
check their length, because we only care about these when we're
serializing the parse tree to a string -- they aren't actual data.